app/requirements.txt
python-decouple==3.8
fastapi==0.115.12
sqlalchemy==2.0.41
alembic==1.15.2
aiosqlite==0.21.0
uvicorn==0.34.2
python-dotenv==1.1.0